Move Version class to its own file
