+++ app/src/main/java/com/github/damontecres/stashapp/Constants.kt
@@ -28,37 +28,6 @@ object Constants {
     const val PREF_KEY_STASH_URL = "stashUrl"
     const val PREF_KEY_STASH_API_KEY = "stashApi"
     const val TAG = "Constants"
-    val MINIMUM_STASH_VERSION = Version("0.23.0")
-}
-
-data class Version(val version: String) {
-    val major: Int
-    val minor: Int
-    val patch: Int
-
-    init {
-        val splits = version.removePrefix("v").split(".").map { it.toInt() }.toList()
-        major = splits[0]
-        minor = splits[1]
-        patch = splits[2]
-    }
-
-    fun isAtLeast(version: Version): Boolean {
-        return this.major > version.major || (
-            this.major == version.major &&
-                (
-                    this.minor > version.minor || this.minor == version.minor &&
-                        this.patch >= version.patch
-                )
-        )
-    }
-}
-
-fun isStashVersionSupported(version: Version?): Boolean {
-    if (version == null) {
-        return false
-    }
-    return version.isAtLeast(Constants.MINIMUM_STASH_VERSION)
 }
 
 /**
